Also move scatter, box, heatmap
